+++ src/oihana/exceptions/http/Error404.php
@@ -7,9 +7,16 @@ use Throwable;
 
 /**
  * Throw an 'not found error' (404).
+ *
+ * The HTTP 404 Not Found client error response status code indicates that the server cannot
+ * find the requested resource. Links that lead to a 404 page are often called
+ * broken or dead links and can be subject to link rot.
+ *
  * @package oihana\exceptions\http
  * @author  Marc Alcaraz (ekameleon)
  * @since   1.0.0
+ *
+ * @see https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Status/404
  */
 class Error404 extends Exception
 {